requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
plotly>=5.13.0
streamlit>=1.28.0
joblib>=1.2.0
seaborn>=0.12.0
Pillow>=9.0.0
mysql-connector-python>=8.0.32
